[breadboard-cli] refactoring
